Apply linting. Remove circular dependency. Move provider components into their respective files
