+++ src/components/enterprise-subsidy-requests/index.js
@@ -1,1 +1,2 @@
-export { default as SubsidyRequestsContextProvider, SubsidyRequestsContext } from './SubsidyRequestsContextProvider';
+export { default as SubsidyRequestsContextProvider } from './SubsidyRequestsContextProvider';
+export { default as SubsidyRequestsContext } from './SubsidyRequestsContext';
